splitted steps into subfolders
