+++ lib/scrape/flow/steps/html/extract_description.ex
@@ -1,16 +1,16 @@
-defmodule Scrape.Flow.Steps.ExtractTitle do
+defmodule Scrape.Flow.Steps.HTML.ExtractDescription do
   @moduledoc false
 
   use Scrape.Flow.Step
 
   def execute(%{dom: dom}, _) when not is_list(dom) and not is_tuple(dom) do
-    {:error, :dom_invalid}
+    fail(:dom_invalid)
   end
 
   def execute(%{dom: dom}, _) do
-    case Scrape.IR.DOM.title(dom) do
-      "" -> assign(title: nil)
-      title -> assign(title: title)
+    case Scrape.IR.DOM.description(dom) do
+      "" -> assign(description: nil)
+      description -> assign(description: description)
     end
   end
