+++ lib/scrape/flow/steps/html/extract_icon_url.ex
@@ -1,4 +1,4 @@
-defmodule Scrape.Flow.Steps.ExtractFeedURLs do
+defmodule Scrape.Flow.Steps.HTML.ExtractIconURL do
   @moduledoc false
 
   use Scrape.Flow.Step
@@ -8,11 +8,11 @@ defmodule Scrape.Flow.Steps.ExtractFeedURLs do
   end
 
   def execute(%{dom: dom, url: url}, _) do
-    assign(feed_urls: Scrape.IR.DOM.feed_urls(dom, url))
+    assign(icon_url: Scrape.IR.DOM.icon_url(dom, url))
   end
 
   def execute(%{dom: dom}, _) do
-    assign(feed_urls: Scrape.IR.DOM.feed_urls(dom))
+    assign(icon_url: Scrape.IR.DOM.icon_url(dom))
   end
 
   def execute(_, _) do
